requirements.txt
langchain
langchain-core
langchain_community
langchain_classic
langchain-huggingface
sentence-transformers
langchain-google-genai
google-generativeai
python-dotenv
streamlit
chromadb
pypdf
pymupdf
unstructured
pdf2image
pdfminer.six
unstructured_pytesseract
unstructured_inference
